Add SettingStore to store user settings, add ValueSetting boilerplate, abstract dividers from setting class
(*) The settings store class will contain a running list of all the profile settings in the app that can be accessed statically and saved/written to the disk or to an external server, as the case may be. Right now I am using the user's hdd, storing the values into the preferences.
(*) ValueSetting will be a setting that stores a generic state value, i.e. a string ValueSetting will be a setting that allows the user to input some text, a boolean ValueSetting will allow the user to toggle a setting on or off